+++ src/ethereum/EthereumNetwork.ts
@@ -270,14 +270,14 @@ export class EthereumNetwork {
   }
 
   start(): void {
-    this.connectNetworkAdapters()
-    this.setupAdapterSubscriptions()
+    // this.connectNetworkAdapters()
+    // this.setupAdapterSubscriptions()
     this.needsLoopTask.start()
   }
 
   stop(): void {
     this.needsLoopTask.stop()
-    this.disconnectNetworkAdapters()
+    // this.disconnectNetworkAdapters()
     // TODO: Abort all in-flight network sync requests
   }
 
@@ -495,6 +495,19 @@ export class EthereumNetwork {
         this.ethEngine.walletLocalData.highestTxBlockHeight,
         highestTxBlockHeight
       )
+
+      // Update the main blockHeight if we received transactions with a higher blockHeight
+      if (highestTxBlockHeight > this.ethEngine.walletLocalData.blockHeight) {
+        this.ethEngine.log(
+          `Updating blockHeight from transactions: ${this.ethEngine.walletLocalData.blockHeight} -> ${highestTxBlockHeight}`
+        )
+        this.ethEngine.checkDroppedTransactionsThrottled()
+        this.ethEngine.walletLocalData.blockHeight = highestTxBlockHeight
+        this.ethEngine.currencyEngineCallbacks.onBlockHeightChanged(
+          this.ethEngine.walletLocalData.blockHeight
+        )
+      }
+
       this.ethEngine.walletLocalDataDirty = true
       this.ethEngine.updateOnAddressesChecked()