+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 4.56.8 (2025-08-04)
+
 - added: (EVM) Accept transactions with a provided `otherParams.nonce` even if pending transactions exist
 - fixed: (EVM) Update blockHeight immediately when receiving transactions to prevent them from appearing as "syncing"